optimized commands; excluded reuseable methods in helper-classes
